airflow/requirements.txt
apache-airflow-providers-google
pandas
chardet
pyarrow